+++ src/Factory/HtmlPurifierConfigFactory.php
@@ -24,26 +24,27 @@ class HtmlPurifierConfigFactory
             include_once $moduleConfig['standalone_path'];
         }
 
-        $config         = isset($moduleConfig['config']) ? $moduleConfig['config'] : [ ];
-        $purifierConfig = self::createConfig($config);
+        $config      = isset($moduleConfig['config']) ? $moduleConfig['config'] : [ ];
+        $definitions = isset($moduleConfig['definitions']) ? $moduleConfig['definitions'] : [ ];
+
+        $purifierConfig = self::createConfig($config, $definitions);
 
         return $purifierConfig;
     }
 
     /**
-     * @param array $config
+     * @param HTMLPurifier_Config $purifierConfig
+     * @param array               $definitions
      *
      * @throws \HTMLPurifier_Exception
      *
      * @return HTMLPurifier_Config
      */
-    public static function createConfig(array $config)
+    public static function createConfig(array $config, array $definitions = [])
     {
         if (isset($config['definitions'])) {
             $definitions  = $config['definitions'];
             unset($config['definitions']);
-        } else {
-            $definitions = [];
         }
 
         $purifierConfig = HTMLPurifier_Config::create($config);
@@ -56,11 +57,37 @@ class HtmlPurifierConfigFactory
                 continue;
             }
 
-            foreach ($methods as $method => $args) {
-                call_user_func_array([ $definition, $method ], $args);
+            foreach ($methods as $method => $invocations) {
+                $invocations = self::convertSingleInvocationToArray($invocations);
+                foreach ($invocations as $args) {
+                    call_user_func_array([ $definition, $method ], $args);
+                }
             }
         }
 
         return $purifierConfig;
+    }
+
+
+    /**
+     * @param array $invocations
+     *
+     * @return array[]
+     */
+    private static function convertSingleInvocationToArray(array $invocations)
+    {
+        if (count($invocations) !== 3) {
+            return $invocations;
+        }
+
+        $allArgumentsAreArray = array_reduce($invocations, function($carry, $value){
+            return is_array($value) && $carry;
+        }, true);
+
+        if ($allArgumentsAreArray) {
+            return $invocations;
+        }
+
+        return [ $invocations ];
     }
 }