+++ tests/Factory/HtmlPurifierConfigFactoryTest.php
@@ -159,6 +159,118 @@ class HtmlPurifierConfigFactoryTest extends TestCase
         $this->assertInstanceOf('HTMLPurifier_AttrDef_Enum', $attributeDefinition);
     }
 
+    public function testFactoryCanSetMultipleDefinitions()
+    {
+        $validAttributes = ['foo','bar','baz','bat'];
+
+        $this->setConfigService([
+            'soflomo_purifier' => [
+                'standalone' => false,
+                'config'     => [
+                    'HTML.DefinitionID'    => 'custom definitions',
+                    'Cache.DefinitionImpl' => null,
+                    'definitions'          => [
+                        'HTML' => [
+                            'addAttribute' => [
+                                [ 'a', 'foo', new HTMLPurifier_AttrDef_Enum($validAttributes) ],
+                                [ 'span', 'bar', 'Bool' ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ]);
+
+        /* @var HTMLPurifier_Config $purifier */
+        $purifierConfig = $this->factory->__invoke($this->serviceManager);
+
+        /** @var HTMLPurifier_HTMLDefinition $definition */
+        $definition = $purifierConfig->getDefinition('HTML');
+        $this->assertInstanceOf('HTMLPurifier_HTMLDefinition', $definition);
+
+        $this->assertInstanceOf('HTMLPurifier_ElementDef', $definition->info['a']);
+        $this->assertInstanceOf('HTMLPurifier_AttrDef_Enum', $definition->info['a']->attr['foo']);
+
+        $this->assertInstanceOf('HTMLPurifier_ElementDef', $definition->info['span']);
+        $this->assertInstanceOf('HTMLPurifier_AttrDef_HTML_Bool', $definition->info['span']->attr['bar']);
+    }
+
+    public function testCanUseDefinitionsAsTopModuleConfigArrayKey()
+    {
+        $validAttributes = ['foo','bar','baz','bat'];
+
+        $this->setConfigService([
+            'soflomo_purifier' => [
+                'standalone' => false,
+                'config'     => [
+                    'HTML.DefinitionID'    => 'custom definitions',
+                    'Cache.DefinitionImpl' => null,
+                ],
+                'definitions'          => [
+                    'HTML' => [
+                        'addAttribute' => [
+                            [ 'a', 'foo', new HTMLPurifier_AttrDef_Enum($validAttributes) ],
+                            [ 'span', 'bar', 'Bool' ],
+                        ],
+                    ],
+                ],
+            ],
+        ]);
+
+        /* @var HTMLPurifier_Config $purifier */
+        $purifierConfig = $this->factory->__invoke($this->serviceManager);
+
+        /** @var HTMLPurifier_HTMLDefinition $definition */
+        $definition = $purifierConfig->getDefinition('HTML');
+        $this->assertInstanceOf('HTMLPurifier_HTMLDefinition', $definition);
+
+        $this->assertInstanceOf('HTMLPurifier_ElementDef', $definition->info['a']);
+        $this->assertInstanceOf('HTMLPurifier_AttrDef_Enum', $definition->info['a']->attr['foo']);
+
+        $this->assertInstanceOf('HTMLPurifier_ElementDef', $definition->info['span']);
+        $this->assertInstanceOf('HTMLPurifier_AttrDef_HTML_Bool', $definition->info['span']->attr['bar']);
+    }
+
+
+
+    public function testAsManyCustomDefinitionsAsDefinitionArgs()
+    {
+        $this->setConfigService([
+            'soflomo_purifier' => [
+                'standalone' => false,
+                'config'     => [
+                    'HTML.DefinitionID'    => 'custom definitions',
+                    'Cache.DefinitionImpl' => null,
+                ],
+                'definitions'          => [
+                    'HTML' => [
+                        'addAttribute' => [
+                            [ 'a', 'bar', 'Bool' ],
+                            [ 'span', 'bar', 'Bool' ],
+                            [ 'div', 'bar', 'Bool' ],
+                        ],
+                    ],
+                ],
+            ],
+        ]);
+
+        /* @var HTMLPurifier_Config $purifier */
+        $purifierConfig = $this->factory->__invoke($this->serviceManager);
+
+        /** @var HTMLPurifier_HTMLDefinition $definition */
+        $definition = $purifierConfig->getDefinition('HTML');
+        $this->assertInstanceOf('HTMLPurifier_HTMLDefinition', $definition);
+
+        $this->assertInstanceOf('HTMLPurifier_ElementDef', $definition->info['a']);
+        $this->assertInstanceOf('HTMLPurifier_AttrDef_HTML_Bool', $definition->info['a']->attr['bar']);
+
+        $this->assertInstanceOf('HTMLPurifier_ElementDef', $definition->info['span']);
+        $this->assertInstanceOf('HTMLPurifier_AttrDef_HTML_Bool', $definition->info['span']->attr['bar']);
+
+        $this->assertInstanceOf('HTMLPurifier_ElementDef', $definition->info['div']);
+        $this->assertInstanceOf('HTMLPurifier_AttrDef_HTML_Bool', $definition->info['div']->attr['bar']);
+    }
+
     protected function setConfigService($array)
     {
         $this->serviceManager->setService('config', $array);